fix(image): 이미지 사이즈 소수점 이하 2자리까지 저장

+++ src/main/java/chzzk/grassdiary/domain/image/entity/Image.java
@@ -24,9 +24,9 @@ public class Image extends BaseCreatedTimeEntity {
 
     private String imageName;
 
-    private Long size;
+    private Double size;
 
-    public Image(String imagePath, String imageName, Long size) {
+    public Image(String imagePath, String imageName, Double size) {
         this.imagePath = imagePath;
         this.imageName = imageName;
         this.size = size;
